requirements.txt
python-dotenv==1.0.0
openai==1.12.0
httpx==0.27.0
aiosmtplib==2.0.2
email-validator==2.1.0.post1
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
aioimaplib==1.0.1
